+++ capturecat-core/src/main/java/com/capturecat/core/api/image/ImageController.java
@@ -1,24 +1,24 @@
-package com.capturecat.core.api.tag;
+package com.capturecat.core.api.image;
 
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.capturecat.core.service.tag.TagService;
+import com.capturecat.core.service.image.ImageService;
 import com.capturecat.core.support.response.ApiResponse;
 
 import lombok.RequiredArgsConstructor;
 
 @RestController
 @RequiredArgsConstructor
-public class TagController {
+public class ImageController {
 
-    private final TagService tagService;
+    private final ImageService imageService;
 
     @PostMapping("/api/v1/images/{imageId}/tags")
     public ApiResponse<?> addTagsToImage(@PathVariable Long imageId, @RequestBody AddTagsToImageRequest request) {
-        tagService.addTagsToImage(imageId, request);
+        imageService.addTagsToImage(imageId, request);
         return ApiResponse.success();
     }
 }
